// src/app/dto/model/user.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User {
    id?: number;
    nom?: string;
    prenom?: string;
    email?: string;
    motDePasse?: string;
    numeroTelephone?: number;
    adresse?: string;
    nouveauteBanque?: Array<string>;
    duree?: number;
    apport?: number;
    mensualite?: number;
    files?: string[]
}